Reunification of ExInfo
- Move the static methods from the ExceptionTracker to the ExInfo
- Remove the ExceptionTracker completely
- Merge the ExceptionTrackerBase into the ExInfo
